Fix incorrect encoding of CMP
We can see the following crash by running Shecc at qemu 5.1.0

	qemu: uncaught target signal 4 (Illegal instruction) - core dumped

The illegal instruction is due to incorrect machine code of CMP
	1d604: e1511002  cmp r1, r2

According to `ARM Architecture Reference Manual ARMv7-A and ARMv7-R`,
the bit<15:12> is 0000, but Shecc encoded as 0001.

The rd in __cmp_r() endcoded as (rd << 12) and, actually,
no one uses rd and disobey the specification.

After Qemu 4.1.0,

https://lore.kernel.org/qemu-devel/20190904193059.26202-1-richard.henderson@linaro.org/

A strict encoding checker was introduced at target/arm/a32.decode,
and the CMP (register) should be:

CMP_xrrr         .... 000 1010 1 .... 0000 .... 0 .. 1 ....
                                      |  |
                                     15  12

That is, the bit<15:12> must be 0000.

Fix this by droping argument rd in __cmp_r().