Enhance find_func through trie structure (#59)
After analyzing with uftrace, it was determined that strcmp() consumes
the most execution time due to the find_ series function, such as
find_func, utilizing strcmp to check for function declarations. At
present, each inserted name is compared every time until a match is
found or it is determined that it has not been inserted. To enhance
search performance, utilizing a trie data structure to store the
inserted names is suggested.

This patch revises the structure of the trie. The 'index' member now
stores the index of the function in the 'FUNCS' array, while the 'next'
member keeps track of the index of the next trie node. A value of 0 for
'index' indicates that the function has not been inserted, and a value
of 0 for 'next' means there is no next trie node. There are 128 'next'
elements, corresponding to the 128 different characters in ASCII.
Although the user may only utilize a subset of these characters, it is
important to perform the necessary checks before insertion.

The implementation has been modified to utilize the original 'FUNCS'
array for storing function information. When adding a new function, the
index is also inserted into the trie. Furthermore, the 'strcmp' function
has been replaced with 'insert_trie', which verifies if the function has
already been inserted.

A sample code has been created to analyze the performance of the
find_func function. The code calculates the sum of the return values of
the function from 1 to n, and the main function returns this sum. Here
is an example of the code where the main function returns the sum of
the return values of f1, f2, and f3.

int f1() { return 1; }
int f2() { return 2; }
int f3() { return 3; }
int main() { return f1() + f2() + f3(); }

Experiments were conducted with find_func, ranging from 1 function to
1024 functions, and the self-time calculated by uftrace is as follows:

|#Function|#find_func called| Before  |      After       |
|---------|-----------------|---------|------------------|
|1        |451              |263.596us|31.329us          |
|2        |456              |296.300us|35.612us          |
|4        |466              |313.824us|38.806us          |
|8        |486              |349.073us|49.190us          |
|16       |526              |413.888us|43.033us          |
|32       |606              |904.453us|64.273us          |
|64       |766              |1.890ms  |95.702us          |
|128      |1086             |4.436ms  |133.466us         |
|256      |1726             |16.381ms |193.744us         |
|512      |3006             |62.848ms |276.330us         |
|1024     |5566             |241.019ms|Segmentation Fault|
---------------------------------------------------------

When the number of functions exceeds 919, the find_func utilizing the
trie structure encounters a Segmentation Fault. This occurs because
only 1024 units were allocated for the trie, which is the same as the
maximum number of functions. As known, the trie structure requires
more than one unit to store a function if the function name has
multiple characters. Consequently, the maximum number of functions
becomes limited by the maximum number of units in the trie, which
depends on the length of the function names.

In spite of the constraint on the maximum number of functions imposed
by the trie structure, a notable enhancement in performance has been
achieved. The magnitude of improvement increases proportionally with the
increase in the number of functions.

Close #7