Eliminate overflow in format manipulation (#63)
The original format will overflow while outputting an integer that is
less than -1e9. Additionally, an unexpected overflow occurs when
flipping the sign while 'val' equals INT_MIN.

The 'fd' output in the printf function was misplaced with 0, which
corresponds to stdin. To fix the output 'fd,' change it to 1, which is
the expected stdout. To ensure that the output is the same as expected,
add a testing function and samples in 'tests/driver.sh'.