Support function-like macro
The implementation redirects the pointer to the source to the beginning
of the body of the macro, like how the copy and paste works. Because of
this, the answer of the second test is 7, not 9.

The '__VA_ARGS__' identifier will expand the remained arguments that
are undeclared. It is useful to write the more readable codes.

Not support nested macro yet.