Refactor the passing of `macro_t`
Both the arguments declared in function and macro can be accessed in
thier children blocks. If we know which function or macro a block
belongs to, we can trace its variables. The simple way to do this is
recording them in the `block_t` structure.

Drop `prev_return_idx` in `macro_t` because we only support non-nested
macro. It used to record the return index of its parent.

Change the prototype of `add_block` in `src/globals.c` to keep the
semantics clear.