Support function-like macro (#68)
The implementation redirects the pointer to the source to the beginning
of the body of the macro, like how the copy and paste works. Because of
this, the answer of the second test is 7, not 9.

The '__VA_ARGS__' identifier will expand the remained arguments that
are undeclared. It is useful to write the more readable codes.

Both the arguments declared in function and macro can be accessed in
their children blocks. If we know which function or macro a block
belongs to, we can trace its variables. The simple way to do this is
recording them in the `block_t` structure.

Nested macros are not supported.