Setup the debug build, update makefiles
Use macro to perform the conditional compilation, and dump the IR in the
debug mode only. Update the guideline in `README.md` also.

Modify the frontend for parsing the function-like macro without body.

The previous `Makefile` does not check the `ARCH` that user typing in.
If it is neither `arm` nor `riscv`, the exception will arise. Fix this
by checking the argument first.

Make the auto-generated templates prettier.