Introduce SSA-based middle end
The current cfront and IR generation are high coupling that making
the flow analysis harder. To introduce the further optimizations,
the static single assignment (SSA) is imperative. This patch
introduces the minimal changes for it, and also keeps the old first
phase IR (ph1_ir) for the compatibility to the upcoming works.