Introduce snapshot test
The current unit test only covers the end to end case. It is hard
the classify which part went wrong in the compilation. The new middle
end can dump the IR with the dot file as a snapshot. We can also use
the tool like `graphviz` to check the flow during the compilation.

Use `make check-snapshots` to invoke the snapshot test, and use
`tests/update-snapshots.sh` to update the snapshots if there's any
change for the cfont.