Peephole: simple instruction fusion
Simplify the generated machine code according to the specification of
ARMv7-A and RV32 ISA.

The `MOV` after arithmetical or `LOAD` instruction is redundant.
Simplily assign the result to the destination register of the `MOV`
instruction.

For example, the generated ARMv7 instructions of the statement
`t = 1 + 2 + 4 + 8` after optimizing are:

```
  mov	r0, 1
  mov	r1, 2
  add	r2, r0, r1
  mov	r0, 4
  add	r1, r2, r0
  mov	r0, 8
- add	r2, r1, r0
- mov	r0, r2
+ add	r0, r1, r0
```

The `r0` register in last two instructions is reused and eliminates the
redundant `MOV` instruction.