CSE: eliminate common load instructions
Reuse the read data if it came from the same location of memory. For
example, the following code references the same element of array:

```
int main()
{
    char arr[1];
    int i, t, pos = 0;
    for (i = 0; i < 10000000; i++)
        t = arr[pos] + arr[pos];
    return 0;
}
```

After the optimization, the common statement `arr[pos]` was substituted
by coping the read result.

```
  ldr	r0, [sp, 4]
  ldr	r1, [sp, 9]
  add	r2, r0, r1
  ldrb	r3, [r2]
  add	r2, r0, r1
- ldrb	r4, [r2]
- add	r3, r3, r4
+ mov	r2, r3
+ add	r2, r3, r2
```

It approximately reduces 5% run time on Raspberry Pi 3B.