Migrate C preprocessor directive handling (#106)
This commit migrates the handling of C preprocessor directives to the
parser unit. This change unifies the behavior of the lexer unit,
enabling it to focus solely on lexical analysis. Additionally, this
enhances the parsing of preprocessor directives by processing them as
tokens instead of raw strings.

Note: In this commit, several temporary hacks have been introduced to
address issues arising from the previous design's shortcomings (see
lexer.c#lex_accept for details).