Fix unsorted slots in snapshot test
The current snapshot is a JSON object which dumped by the `dot`
utility. The keys of object or the elements of arrays might not be
sorted by the `jq` and could make the `diff` command failed. The
incoming changes make sure the keys and the elements to be sorted
before saving/checking the snapshots.