Fix bad logical-and implementation
The previous implementation of the logical-and operation produced
incorrect result. This implementation ensures the correctness of the
logical-and behavior and guarantees left-to-right evaluation as
mentioned in the C99 standard.

Add new function, `read_log_and_operation` and `finalize_log_and`.

`read_log_and_operation` is used when `read_expr` encounters
`OP_log_and`. This function creates a new branch instruction to test the
operand before `OP_log_and`, and then creates a new basic block for next
expression.

`finalize_log_and` is used when all operands are consumed or when an
operator with lower priority than logical-and is encountered. This
function create new basic blocks and new branchs to derive the final
value (0 or 1) for the result of the logical-and operation, and then
creates a new basic block for other new instructions.

In `read_expr`, we provide data for the false branch of logical-and
operation, including `else_bb` and `land_else_label`. All operands of
logical-and operator are designed to branch to the same "else" basic
block for the false branch.

This commit adjusts the order of connecting the basic block during
parsing `T_for` and `T_while`. This change ensures that basic blocks are
connected correctly when the condition expression includes a logical-and
operation.

Resolve #137