CI: Fix error when running on host-arm
When pulling a base image from Docker Hub, the architecture defaults to
that of the native host, which is typically x86_64 or amd64 on GitHub
runners. Since the run-on-arch-action GitHub Action aims to run shecc on
an Armv7 Ubuntu image, it attempts to pull the native host manifest
(either x86_64 or amd64), resulting in a 'no match for platform in
manifest' error unless the --platform option is specified. Therefore,
specifying the --platform option as linux/armv7 resolves this issue.