Ensure the correctness when using formatted output conversion
The implementation of __format() produces incorrect results
when using field width and flag characters.

For example, consider the following printf calls, where the
prepended characters are placed in incorrect positions.

    printf("%#12x\n", 0xff00cde1) --> "0x  ff00cde1\n"
    printf("%#12o\n", 0x1000c)    --> "0     200014\n"
    printf("%12d\n", -3580)	  --> "0000000-3580\n"

Thus, this commit improves __format() to ensure correctness
and to make the results consistent with GCC, and adds a test
case to validate the changes.