Correct the return value of fputc function
According to the man page, the return value of fputc() should be either
the character written as an "unsigned char" cast to an "int" or EOF on
error, but the current implementation doesn't follow the specification.

Therefore, these changes correct the implementation to return the ASCII
code of the written character or -1 if the output fails, thereby
matching the description in the man page.

Additionally, the necessary test cases are also added to validate the
correctness.