Use explicit "void" parameter instead
Adds the keyword "void" to the parameter lists of functions.  This helps
validation of the implementation of "void" parameter lists in the
bootstrapping process.  Especially K&R style exists until C23, though
it's not implemented in shecc now, shecc is first built by another
compiler such as gcc and is written in C99 (as per Makefile).

Suggested-by: Yu-En Hsiao <drxiao6399@gmail.com>