Fix the behavior of local array initializer
After observing the implementation of cfront, it appears that the local
array initializer may not generate correct instructions to initialize
all elements.

For example, consider the following code:

    int main()
    {
        int a[5] = {5, 10};
        return a[0] + a[1] + a[2] + a[3] + a[4]
    }

If the number of elements in the initializer-list is less than the size
of array, the generated instructions only initialize the first elements.
The remaining elements are left uninitialized, which may cause the
return value in the above example to differ from the expected 15.

When the initializer-list is empty, all elements remain uninitialized.

Therefore, this commit improves the cfront to generate more instructions
to initialize all elements of an array exactly.