Detect the sysroot path of the ARM GNU toolchain automatically
Originally, the default path of ld-linux.so is set to a specific path
for Arm architecture. However, Considering that developers may manually
install the ARM GNU toolchain, the sysroot path may not match the
default path.

This commit improves arm.mk to use a more convenient method to
automatically detect the correct sysroot path.