Ensure stack is always aligned with 8 bytes
In dynamic linking mode, consider external functions may use data types
such as 'long long int' or other equivalents. The program may encounter
a Bus error if external functions access 8-byte data on the stack that
is not properly aligned.

To prevent the aforementioned issue, these changes adjust the Arm code
generator to ensure that the program's stack is always aligned with 8
bytes.

Since the RISC-V architecture does not yet support dynamic linking, the
corresponding code generator is not modified in this commit.